File Input and Traversal
Now the input is read from a file. Also added the 3 traversals. They are
stored in ArrayLists for easier navigation later or to print in order.
uses set and get functions. all traversals are recursive. Also tested an
works as intended. Need to look at outputing to a file, printing the
tree itself, and changing how the program handles duplicate entries as
per the assignment.